+++ ios/RCTVideoE.m
@@ -1,5 +1,5 @@
 #import <React/RCTConvert.h>
-#import "RCTVideo.h"
+#import "RCTVideoE.h"
 #import <React/RCTBridgeModule.h>
 #import <React/RCTEventDispatcher.h>
 #import <React/UIView+React.h>
