+++ ios/RCTVideoEPlayerViewControllerDelegate.h
