+++ ios/RCTVideoManagerE.h
